WIP : GUI Editor
https://trello.com/c/CgePN3br
Refactor light component property editor widgets.

+++ Source/Editor/InnocenceEditor/combolabeltext.cpp
@@ -2,58 +2,58 @@
 
 ComboLabelText::ComboLabelText(QWidget *parent) : QWidget(parent)
 {
-    m_label = nullptr;
-    m_text = nullptr;
+	m_label = nullptr;
+	m_text = nullptr;
 }
 
 void ComboLabelText::Initialize(QString labelText)
 {
-    m_validator = new QDoubleValidator(-2147483647, 2147483647, 4);
-    m_validator->setProperty("notation", QDoubleValidator::StandardNotation);
+	m_validator = new QDoubleValidator(-2147483647, 2147483647, 4);
+	m_validator->setProperty("notation", QDoubleValidator::StandardNotation);
 
-    m_text = new QLineEdit();
-    m_text->setValidator(m_validator);
+	m_text = new QLineEdit();
+	m_text->setValidator(m_validator);
 
-    m_label = new AdjustLabel();
-    m_label->setText(labelText);
-    m_label->setAlignment(Qt::AlignRight);
-    m_label->Initialize(m_text);
+	m_label = new AdjustLabel();
+	m_label->setText(labelText);
+	m_label->setAlignment(Qt::AlignRight);
+	m_label->Initialize(m_text);
 
-    connect(m_label, SIGNAL(Adjusted()), this, SLOT(TextGotEdited()));
-    connect(m_text, SIGNAL(textEdited(QString)), this, SLOT(TextGotEdited()));
+	connect(m_label, SIGNAL(Adjusted()), this, SLOT(TextGotEdited()));
+	connect(m_text, SIGNAL(textEdited(QString)), this, SLOT(TextGotEdited()));
 }
 
 void ComboLabelText::AlignLabelToTheLeft()
 {
-    m_label->setAlignment(Qt::AlignLeft | Qt::AlignVCenter);
+	m_label->setAlignment(Qt::AlignLeft | Qt::AlignVCenter);
 }
 
 void ComboLabelText::AlignLabelToTheRight()
 {
-    m_label->setAlignment(Qt::AlignRight | Qt::AlignVCenter);
+	m_label->setAlignment(Qt::AlignRight | Qt::AlignVCenter);
 }
 
 AdjustLabel*ComboLabelText::GetLabelWidget()
 {
-    return m_label;
+	return m_label;
 }
 
 QLineEdit*ComboLabelText::GetTextWidget()
 {
-    return m_text;
+	return m_text;
 }
 
 float ComboLabelText::GetAsFloat()
 {
-    return m_text->text().toFloat();
+	return m_text->text().toFloat();
 }
 
 void ComboLabelText::SetFromFloat(float value)
 {
-    m_text->setText(QString::number(value));
+	m_text->setText(QString::number((double)value));
 }
 
 void ComboLabelText::TextGotEdited()
 {
-    emit ValueChanged();
+	emit ValueChanged();
 }
